Finalize SLP driver (Issue #58)
This fixes the margin (centering) of the label image and sorts the label sizes
so that it is easier to find the size.  Note, however, that the advertised size
reflects the printable area (per the CUPS driver usage) and not the marketing
size, e.g., the 36x89mm large shipping label is reported as 35x83mm - half a mm
on the sides and 3mm top/bottom for margins.  I might decide to change that if
there is enough interest...

Note: These printers don't support status reporting of any kind despite having
a status command defined (0x01).  The USB Get-Status values are also wrong -
the printer always says its cover is open and it is out of media/labels...