Revert DYMO label margins to before ed195b5233
When DYMO tape printing support was added in ed195b5233 and 6f49c6e066,
the label left/right margin was changed from 100 to 367 and the
bottom/top margin from 525 to 1.  This causes the label content to be
offset both vertically and horizontally with reference to the label.

The margin values are stored in the state file along with the other
media details so it is necessary to re-select the media size (or remove
the state file) for this change to take effect.